requirements.txt
astor==0.8.1
brotlipy==0.7.0
certifi==2021.10.8
dlib==19.23.1
fonttools==4.25.0
ftfy==6.1.1
gast==0.2.2
mkl-fft==1.3.1
mkl-service==2.4.0
munkres==1.1.4
numpy==1.21.6
opencv-python==4.5.5.64
Pillow==9.1.0
protobuf==3.19.1
regex==2022.4.24
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tqdm==4.64.0
typing_extensions==4.2.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.1